+++ quint/src/simulation.ts
@@ -62,6 +62,8 @@ function errSimulationResult(status: SimulatorResultStatus, errors: ErrorMessage
  *
  * @param idGen a unique generator of identifiers
  * @param code the source code of the modules
+ * @param mainStart the start index of the main module in the code
+ * @param mainEnd the end index of the main module in the code
  * @param mainName the module that should be used as a state machine
  * @param mainPath the lookup path that was used to retrieve the main module
  * @param options simulator settings
@@ -71,6 +73,8 @@ function errSimulationResult(status: SimulatorResultStatus, errors: ErrorMessage
 export function compileAndRun(
   idGen: IdGenerator,
   code: string,
+  mainStart: number,
+  mainEnd: number,
   mainName: string,
   mainPath: SourceLookupPath,
   options: SimulatorOptions
@@ -82,23 +86,22 @@ export function compileAndRun(
   // that are required by the runner.
   // This code should be revisited in #618.
   const o = options
-  const wrappedCode = `${code}
+  // Defs required by the simulator, to be added to the main module before compilation
+  const extraDefs = [
+    `val ${lastTraceName} = [];`,
+    `def q::test(q::nrunsArg, q::nstepsArg, q::initArg, q::nextArg, q::invArg) = false`,
+    `action q::init = { ${o.init} }`,
+    `action q::step = { ${o.step} }`,
+    `val q::inv = { ${o.invariant} }`,
+    `val q::runResult = q::test(${o.maxSamples}, ${o.maxSteps}, q::init, q::step, q::inv)`,
+  ]
 
-module __run__ {
-  import ${mainName}.*
-
-  val ${lastTraceName} = [];
-  def q::test(q::nrunsArg, q::nstepsArg, q::initArg, q::nextArg, q::invArg) = false;
-  action q::init = { ${o.init} }
-  action q::step = { ${o.step} }
-  val q::inv = { ${o.invariant} }
-  val q::runResult =
-    q::test(${o.maxSamples}, ${o.maxSteps}, q::init, q::step, q::inv)
-}
-`
+  // Construct the modules' code, adding the extra definitions to the main module
+  const newMainModuleCode = code.slice(mainStart, mainEnd - 1) + extraDefs.join('\n')
+  const codeWithExtraDefs = code.slice(0, mainStart) + newMainModuleCode + code.slice(mainEnd)
 
   const recorder = newTraceRecorder(options.verbosity, options.rng)
-  const ctx = compileFromCode(idGen, wrappedCode, '__run__', mainPath, recorder, options.rng.next)
+  const ctx = compileFromCode(idGen, codeWithExtraDefs, mainName, mainPath, recorder, options.rng.next)
 
   if (ctx.compileErrors.length > 0 || ctx.syntaxErrors.length > 0 || ctx.analysisErrors.length > 0) {
     const errors = ctx.syntaxErrors.concat(ctx.analysisErrors).concat(ctx.compileErrors)